==> FinalModel/Scripts/02_GetImages.ipynb <==
# Download Images

Download Images to the FileSystem from the Images csv

import pandas as pd
import requests
pd.set_option('max_colwidth', -1)
import time

images = pd.read_csv('../Data/Images.csv', sep = ';')
images.head(5)

images['image'][5:10]

## Save images

#### WARNING: Expensive time to Download
1 min 50 sec = 500 images

flag = 0 #If you want to download a subset change this flag to the Number of images you want to download
url_errors = []

print ('Downloading images...')

for idx, row in images.iterrows():
    
    if flag <= idx:
        if idx%500==0:
            print (str(idx) + ' images downloaded')

        name = row['id'] + '.jpg'

        with open('../Data/Images/'+name, 'wb') as img_file:
            try:
                response = requests.get(row['image'], stream=True)

                if response.ok:

                    for batch in response.iter_content(1024):
                        if not batch:
                            break

                        img_file.write(batch)
                else:
                    print(f'Status code error with {response.status_code} code.')

            except Exception as e:
                print(f'Exception: {e}')
                if url not in url_errors:
                    url_errors.append(row['image'])
                time.sleep(3)
                pass
