==> FinalModel/Scripts/03_TextBuilding.ipynb <==
# Text Building

Build one text column oining description and categories

import pandas as pd
import ast
import os

pd.set_option('max_colwidth', 250)

path = os.path.join('../Data/')
print (os.listdir(path + 'FinalItems/'))

data = pd.read_csv(path + '204kProducts.csv', sep = ';')
data.head(3)

## Join text

df_text = pd.DataFrame(data['alternative_id'])
subset = data[['category', 'name']]

### Category

for i, lista in enumerate(subset['category'].values):
    subset['category'].values[i] = ast.literal_eval(lista)
subset['category'] = subset['category'].apply(lambda x: ' '.join(x))
subset.head(3)

subset.loc[43000]

### Brand

#data['brand'] = data['brand'].str.replace(" ","_")
data.head(3)

### Final Column

df_text['text'] = subset.apply(lambda x: ' '.join(x.dropna().astype(str)), axis = 1)
df_text.head(3)

finaldf = pd.concat([df_text, data['brand']], axis = 1)
finaldf = finaldf[['alternative_id', 'brand', 'text']]
finaldf.head(3)

finaldf.to_csv(path + 'Descriptions204k.csv', sep = ';', index=False)
