==> FinalModel/Scripts/04_DataFiltered.ipynb <==
# Data Filtered

In this notebook I generate the final csv in which i select the columns that I want to keep and drop the ones I don´t

import pandas as pd
import os

path_input = os.path.join('../Data/')
path_output = os.path.join('../Data/FinalItems/')

print(os.listdir(path_input + 'FinalItems/'))

data = pd.read_csv(path_input + '/204kProducts.csv', sep = ';')
data.columns

data.head(1)

columns = ['alternative_id', 'availability_date', 'brand', 'category', 'name', 'price']

data = data[columns]

data.columns = ['item_id', 'availability_date', 'brand', 'category', 'name', 'price']

data.to_csv(path_output + 'data_filtered.csv', index = False, sep = ';')
