==> FinalModel/Scripts/07_TextAutoencoder.ipynb <==
# Text Autoencoder

In this notebook we generate the text autoencoder in order to reduce the dimension of our text vector

import pandas as pd
import os
import numpy as np
import nltk.data
from nltk import word_tokenize, sent_tokenize
from sklearn.feature_extraction.text import CountVectorizer
from nltk.corpus import stopwords
import gensim
from gensim.models import Word2Vec

from keras.preprocessing.sequence import pad_sequences
from keras.preprocessing.text import Tokenizer
from keras.layers import Embedding
from keras.models import Model, Sequential, load_model
from keras.layers import Input, Flatten, Dense, Conv1D, MaxPooling1D, GlobalMaxPool1D, SpatialDropout1D, \
                          UpSampling1D, LSTM, RepeatVector, TimeDistributed, GRU, Bidirectional, concatenate, \
                          GlobalAveragePooling1D
from keras.utils import plot_model, to_categorical
from sklearn.neighbors import NearestNeighbors
from keras.utils.vis_utils import plot_model

from keras.models import model_from_json

pd.set_option('max_colwidth', 250)

path = os.path.join('../Data/')
path_models = os.path.join('../Models/')

data = pd.read_csv(path + 'Texto_PreProcesado.csv', sep = ';', index_col = False)
data.head()

## Vectorize Sentences

- Initialize tokenizer with num_words = MAX_NB_WORDS (200K). i.e. The tokenizer will perform a word count, sorted by number of occurences in descending order and pick top N words, 200K in this case 
- Use tokenizer's texts_to_sequences method to convert text to array of integers.
- The arrays obtained from previous step might not be of uniform length, use pad_sequences method to obtain arrays with length equal to MAX_SEQUENCE_LENGTH (30)

MAX_NB_WORDS = 30_000 #decided by cumsum wordcount plot (Script 01)
MAX_SEQUENCE_LENGTH = 24 #decided by max words in a product (Script 00)
EMBEDDING_DIM = 100 #Same dim as our W2V embedding

data['text'] = data['text'].astype(str)
all_text = data['text']
all_text = all_text.drop_duplicates (keep = False)

tokenizer = Tokenizer(num_words=MAX_NB_WORDS, )
tokenizer.fit_on_texts(all_text)

data_sequences = tokenizer.texts_to_sequences(data['text'])
data_vec = pad_sequences(data_sequences, maxlen=MAX_SEQUENCE_LENGTH)

word_index = tokenizer.word_index
print('Found %s unique tokens.' % len(word_index))

data_vec[0]

#### Lets load our Custom Embedding

modelWV = Word2Vec.load(path_models + "word2vec_model")

#### Build Keras Embedding 

word_vectors = modelWV.wv
vocabulary_size = len(word_index) + 1
embedding_matrix = np.zeros((len(word_index) + 1, EMBEDDING_DIM))

for word, i in word_index.items():
    if word in modelWV:
        embedding_matrix[i] = modelWV[word]
    else:
        embedding_matrix[i] = np.random.rand(1, EMBEDDING_DIM)[0]
            

del(word_vectors)

embedding_layer = Embedding(input_dim = vocabulary_size,
                            output_dim = EMBEDDING_DIM,
                            input_length = MAX_SEQUENCE_LENGTH,
                            weights=[embedding_matrix],
                            name='w2v_embedding',
                            trainable=False)

model = Sequential()
model.add(embedding_layer)
model.compile('adam', 'mse')
model.summary()

data_vec.shape

data_embedded = model.predict(data_vec, verbose = 1)

## Model 1

Dense Layer based, low quantity of params and easy to train

input_i = Input(shape=(MAX_SEQUENCE_LENGTH, ))

text_embedding = Embedding(*embedding_matrix.shape, weights = [embedding_matrix], trainable = False)(input_i)

encoded_h1 = Dense(128, activation='relu')(text_embedding)
encoded_h2 = Dense(64, activation='relu')(encoded_h1)
encoded_h3 = Dense(32, activation='relu')(encoded_h2)
encoded_h4 = Dense(16, activation='relu')(encoded_h3)
#encoded_h5 = Dense(8, activation='relu')(encoded_h4)

latent = Dense(8, activation='relu', name = 'ENCODER')(encoded_h4)

#decoder_h1 = Dense(8, activation='relu')(latent)
decoder_h2 = Dense(16, activation='relu')(latent)
decoder_h3 = Dense(32, activation='relu')(decoder_h2)
decoder_h4 = Dense(64, activation='relu')(decoder_h3)
decoder_h5 = Dense(128, activation='relu')(decoder_h4)

#output = Dense(EMBEDDING_DIM, activation='relu')(decoder_h5)
glob = Flatten()(decoder_h5)
output = Dense(MAX_SEQUENCE_LENGTH, activation = 'relu')(glob)

autoencoder = Model(input_i, output)

autoencoder.compile(optimizer = 'rmsprop', loss = 'mse', metrics = ['acc'])
autoencoder.summary()

%%time
autoencoder.fit(data_vec, data_vec, epochs=10,
            batch_size=128, verbose = 1)

## Model 2

### The more complex the model is the better?

- Basic LSTM Layer based

- Take nothe that we are compressing more the encoder layer, instead of LENTGH = 192 in dense approach, now we are compressing to LENGHT = 128.


input_i = Input(shape=(MAX_SEQUENCE_LENGTH, ))

text_embedding = Embedding(*embedding_matrix.shape, weights = [embedding_matrix], trainable = False)(input_i)

x1 = LSTM(128, return_sequences=False, name = 'ENCODER')(text_embedding)
x2 = RepeatVector(24)(x1)
x3 = LSTM (24, return_sequences=False)(x2)

autoencoder = Model(inputs = input_i, outputs = x3)
autoencoder.compile(optimizer = 'rmsprop', loss = 'mse' , metrics = ['acc'])
autoencoder.summary()

%%time
autoencoder.fit(data_vec, data_vec ,epochs=3,
            batch_size=64, verbose = 1)

## Model 3

- Lets give it another chance to LSTM Layer based but now we will give the LSTM 2 inputs in order to fit better to the order word changes.

- Take nothe that we are compressing more the encoder layer, instead of LENTGH = 192 in dense approach, now we are compressing to LENGHT = 128 as we did in Model 2

- Also we have less training params so we will train the autoencoder faster than in Model2 Approach


input_i1 = Input(shape=(MAX_SEQUENCE_LENGTH, ))
input_i2 = Input(shape=(MAX_SEQUENCE_LENGTH, ))

text_embedding1 = Embedding(*embedding_matrix.shape, weights = [embedding_matrix], trainable = False)(input_i1)
text_embedding2 = Embedding(*embedding_matrix.shape, weights = [embedding_matrix], trainable = False)(input_i2)

am1 = LSTM(64, return_sequences=False)(text_embedding1)
am2 = LSTM(64, return_sequences=False)(text_embedding2)

decoder = concatenate([am1, am2], name = 'ENCODER')

x2 = RepeatVector(24)(decoder)
x3 = LSTM(24, return_sequences=False)(x2)

autoencoder = Model([input_i1, input_i2], x3)

autoencoder.compile(optimizer = 'adam', loss = 'mse', metrics = ['acc'])
autoencoder.summary()

%%time
autoencoder.fit([data_vec, data_vec], data_vec ,epochs=3,
            batch_size=64, verbose = 1)

encoder = Model(inputs=autoencoder.input, outputs=autoencoder.get_layer('ENCODER').output)

encoder.save(path_models + 'encoder_text.h5')

encoder = load_model(path_models + 'encoder_text.h5')

shape = encoder.get_layer('ENCODER').output_shape[1] * encoder.get_layer('ENCODER').output_shape[2]
shape

encoder.summary()

### Most similar products

data.loc[70000]

query = data_vec[70000]
query.shape

X_test = data_vec.copy()
X_test.shape

codes = encoder.predict(X_test)
codes.shape

codes.shape

query_code = encoder.predict([query.reshape(1, MAX_SEQUENCE_LENGTH)])
query_code.shape

codes = codes.reshape(-1, shape)
print(codes.shape)
query_code = query_code.reshape(1, shape)
print(query_code.shape)

### Fit the KNN to the test set

n_neigh = 10
nbrs = NearestNeighbors(n_neighbors=n_neigh).fit(codes)

distances, indices = nbrs.kneighbors(np.array(query_code))

closest_sent = X_test[indices]
print(closest_sent.shape)
closest_sent = closest_sent.reshape(n_neigh, MAX_SEQUENCE_LENGTH); 
print(closest_sent.shape)

data.loc[58000]

mis_indices = indices.tolist()[0]
for i in range(n_neigh):
    print (data.loc[mis_indices[i]])
    print('-'*50)
