==> FinalModel/Scripts/10_RankModel.ipynb <==
# Rank Model

This model will rank the items generated by the candidates model and sort recommendations due to the ranked output. Trys to predict the quantity of times the customer is going to purchase that item.

import pandas as pd
import os, sys
import numpy as np
import seaborn as sns
import gc
import warnings

from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.layers import Embedding
from gensim.models import Word2Vec

from tensorflow.keras.regularizers import l1, l2
from tensorflow.keras.models import Model, Sequential, load_model
from tensorflow.keras.layers import Input, Flatten, Dense, Conv1D, MaxPooling1D, GlobalMaxPool1D, SpatialDropout1D, \
                          LSTM, GRU, concatenate, Bidirectional, \
                          Reshape, Dropout, GlobalAveragePooling1D

from sklearn.preprocessing import StandardScaler, MinMaxScaler

from tensorflow.keras.utils import to_categorical

from sklearn.model_selection import train_test_split
import matplotlib.pyplot as plt

from custom_functions import norm_text, norm_brands, norm_images

pd.set_option('max_colwidth', 250)
warnings.filterwarnings(action='once')

## Load Data

path = os.path.join('../Data/')
path_models = os.path.join('../Models/')

data = pd.read_csv(path + 'FinalItems/data_filtered.csv', sep = ';')
data_processed = pd.read_csv(path + 'FinalItems/data_final.csv')
data_processed['text'] = data_processed['text'].astype(str)

data_processed.head(3)

items_unique = data_processed.item_id.unique()
items_map = {i:val for i,val in enumerate(items_unique)}
items_map_inv = {val:i for i,val in enumerate(items_unique)}
items_map_text = data_processed.set_index('item_id_int').text.to_dict()

customers_unique = data_processed.customer_id.unique()
items_unique = data_processed.item_id.unique()
brand_unique = data_processed.brand.unique()

MAX_NB_WORDS = 30_000 #decided by cumsum wordcount plot (Script 01)
MAX_SEQUENCE_LENGTH = 24 #decided by max words in a product (Script 00)
EMBEDDING_DIM = 100 #Same dim as our W2V embedding

all_text = data_processed['text']
all_text = all_text.drop_duplicates (keep = False)

tokenizer = Tokenizer(num_words=MAX_NB_WORDS, )
tokenizer.fit_on_texts(all_text)

data_sequences = tokenizer.texts_to_sequences(data_processed['text'])
data_vec = pad_sequences(data_sequences, maxlen=MAX_SEQUENCE_LENGTH)

word_index = tokenizer.word_index
print('Found %s unique tokens.' % len(word_index))

### Load Embedding

modelWV = Word2Vec.load(path_models + 'word2vec_model')

word_vectors = modelWV.wv
vocabulary_size = len(word_index) + 1
embedding_matrix = np.zeros((len(word_index) + 1, EMBEDDING_DIM))

for word, i in word_index.items():
    if word in modelWV:
        embedding_matrix[i] = modelWV[word]
    else:
        embedding_matrix[i] = np.random.rand(1, EMBEDDING_DIM)[0]
            

del(word_vectors)

## Build rank model

scaler = MinMaxScaler()
data_to_scale = data_processed[['PRICE','score', 'item_age']]
scaled_data = data_to_scale.values

def get_rank_model():
    user_id_input = Input(shape=[1], name='user')
    item_id_input = Input(shape=[1], name='item')
    brand_id_input = Input(shape = [1], name = 'brand')

    price_input = Input(shape = [1], name = 'price')
    item_age_input = Input(shape = [1], name = 'item_age')

    price_power_input = Input(shape = [1], name = 'power_price')
    item_age_power_input = Input(shape = [1], name = 'power_item_age')

    price_sqrt_input = Input(shape = [1], name = 'sqrt_price')
    item_age_sqrt_input = Input(shape = [1], name = 'sqrt_item_age')
    
    sequence_input = Input(shape = (MAX_SEQUENCE_LENGTH, ), name = 'text')

    embedding_size = 20
    item_embedding_size = 50

    user_embedding = Embedding(output_dim=embedding_size, input_dim = customers_unique.shape[0]+1,
                               input_length=1, name='user_embedding')(user_id_input)
    item_embedding = Embedding(output_dim=item_embedding_size, input_dim = items_unique.shape[0]+1,
                               input_length=1, name='item_embedding')(item_id_input)
    brand_embedding = Embedding(output_dim=embedding_size, input_dim = brand_unique.shape[0]+1,
                               input_length=1, name='brand_embedding')(brand_id_input)
    
    text_embedding = Embedding(*embedding_matrix.shape, weights = [embedding_matrix], trainable = False)(sequence_input)

    user_vecs = Reshape([embedding_size])(user_embedding)
    item_vecs = Reshape([item_embedding_size])(item_embedding)
    brand_vecs = Reshape([embedding_size])(brand_embedding)
    text_flat = Flatten()(text_embedding)

    x1 = concatenate([user_vecs, item_vecs, brand_vecs, price_input, item_age_input,
                     price_power_input, item_age_power_input,
                     price_sqrt_input, item_age_sqrt_input], name = 'VECTOR_PRODUCTO')

    x2 = Dropout(0.5)(x1)
    x3 = Dense (512, activation = 'relu', activity_regularizer=l1(0.0001))(x2)
    x4 = Dropout(0.4)(x3)
    x5 = Dense(256, activation = 'relu', activity_regularizer=l1(0.0001))(x4)
    x6 = Dropout(0.3)(x5)
    x7 = Dense(128, activation = 'relu',activity_regularizer=l1(0.0001))(x6)
    x8 = Dropout(0.2)(x7)

    output = Dense(1, activation = 'linear' )(x8)

    model = Model(inputs=[user_id_input, item_id_input, brand_id_input, price_input, item_age_input,
                         price_power_input, item_age_power_input,
                         price_sqrt_input, item_age_sqrt_input], 
                  outputs = output)

    model.compile(loss = 'mse',
                  optimizer = 'rmsprop',
                  metrics = ['acc'])

    model.summary()
    
    return model

rank_model = get_rank_model()

history = rank_model.fit([data_processed['customer_id_int'], data_processed['item_id_int'], 
                     data_processed['brand_id'], data_processed['PRICE'],
                     data_processed['item_age'],
                     data_processed['power_price'], data_processed['power_item_age'],
                     data_processed['sqrt_price'], data_processed['sqrt_item_age'], data_vec],
                     data_processed['score_original'],
                     epochs = 15,
                     batch_size = 256, 
                     shuffle = True)

rank_model.save(path_models + 'rank_model')

### Lets test our rankings

preds = rank_model.predict([data_processed['customer_id_int'], data_processed['item_id_int'], 
                     data_processed['brand_id'], data_processed['PRICE'],
                     data_processed['item_age'],
                     data_processed['power_price'], data_processed['power_item_age'],
                     data_processed['sqrt_price'], data_processed['sqrt_item_age'], data_vec], verbose = 1)

preds.shape

preds[:10]

data_processed['score_original'][:10]

#### Query ranking

We will take customer 128 as query to give an example of our rankings. This customer is a Harry Potter freak as I show below.

c_idx = data_processed.index[data_processed['customer_id_int'] == 128] 
data_processed[data_processed['customer_id_int'] == 128]

preds[c_idx]

### Conclussion

- We are going to reject this model because is performing a huge bias with the items that are most usually bought as clothes, so it will trend to rank with higher predictions that kind of items even though our customer is a harry potter freak, as we saw in the example above.

- In order to perform a good ranking we will need the info of when a customer is going to buy again?

---
